Move a bunch of IIIF-related components/functions into the catalogue app
They're only used there -- let's shake them (and their dependencies) out
of the common namespace, to keep that nice and lean.
