+++ catalogue/webapp/__mocks__/iiif-manifest.js
