+++ catalogue/webapp/components/IIIFViewer/IIIFCanvasThumbnail.tsx
@@ -5,10 +5,7 @@ import styled from 'styled-components';
 import { iiifImageTemplate } from '@weco/common/utils/convert-image-uri';
 import IIIFResponsiveImage from '@weco/common/views/components/IIIFResponsiveImage/IIIFResponsiveImage';
 import LL from '@weco/common/views/components/styled/LL';
-import {
-  isImageRestricted,
-  getThumbnailService,
-} from '@weco/common/utils/iiif';
+import { isImageRestricted, getThumbnailService } from '../../utils/iiif';
 import Padlock from '@weco/common/views/components/styled/Padlock';
 import Space from '@weco/common/views/components/styled/Space';
