+++ catalogue/webapp/components/IIIFViewer/MainViewer.tsx
@@ -24,7 +24,7 @@ import {
   getImageAuthService,
   isImageRestricted,
   getThumbnailService,
-} from '@weco/common/utils/iiif';
+} from '../../utils/iiif';
 import { font } from '@weco/common/utils/classnames';
 import { IIIFCanvas, SearchResults } from '@weco/common/model/iiif';
 import ItemViewerContext from '@weco/common/views/components/ItemViewerContext/ItemViewerContext';
