+++ catalogue/webapp/components/IIIFViewer/NoScriptViewer.tsx
@@ -1,7 +1,7 @@
 import NextLink from 'next/link';
 import styled from 'styled-components';
 import { classNames } from '@weco/common/utils/classnames';
-import { getServiceId } from '@weco/common/utils/iiif';
+import { getServiceId } from '../../utils/iiif';
 import IIIFResponsiveImage from '@weco/common/views/components/IIIFResponsiveImage/IIIFResponsiveImage';
 import { iiifImageTemplate } from '@weco/common/utils/convert-image-uri';
 import { imageSizes } from '@weco/common/utils/image-sizes';
