+++ catalogue/webapp/package.json
@@ -17,6 +17,7 @@
     "@types/jest": "^26.0.17",
     "@types/koa-router": "^7.4.4",
     "@types/react": "^16.9.35",
+    "@types/react-window": "^1.8.2",
     "@types/supertest": "^2.0.11",
     "@weco/common": "1.0.0",
     "@weco/toggles": "1.0.0",
@@ -34,6 +35,7 @@
     "koa": "^2.5.0",
     "koa-router": "^7.4.0",
     "lazysizes": "^5.2.1",
+    "lodash.clonedeep": "^4.5.0",
     "next": "^11.1.2",
     "nodemon": "^1.17.5",
     "raven-js": "^3.27.0",
