+++ catalogue/webapp/pages/api/work.ts
@@ -3,7 +3,7 @@ import {
   getAudio,
   getVideo,
   getDownloadOptionsFromManifest,
-} from '@weco/common/utils/iiif';
+} from '../utils/iiif';
 import { NextApiRequest, NextApiResponse } from 'next';
 
 export default async (
