+++ catalogue/webapp/utils/iiif.ts
