+++ common/package.json
@@ -26,7 +26,6 @@
     "@types/jest": "^26.0.17",
     "@types/lodash.debounce": "^4.0.6",
     "@types/lodash.groupby": "^4.6.6",
-    "@types/react-window": "^1.8.2",
     "@types/uuid": "^8.3.0",
     "@weco/toggles": "1.0.0",
     "@wojtekmaj/enzyme-adapter-react-17": "^0.6.3",
@@ -44,7 +43,6 @@
     "jest": "^26.6.3",
     "koa-compose": "^4.1.0",
     "lazysizes": "^5.2.1",
-    "lodash.clonedeep": "^4.5.0",
     "lodash.debounce": "^4.0.8",
     "lodash.flattendeep": "^4.4.0",
     "lodash.groupby": "^4.6.0",
